requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
matplotlib>=3.7.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
albumentations>=1.3.0
seaborn>=0.12.0
